Sign conversion fixes

+++ src/core/client_config/connector.h
@@ -50,7 +50,7 @@ typedef struct {
   grpc_pollset_set *interested_parties;
   /** address to connect to */
   const struct sockaddr *addr;
-  int addr_len;
+  size_t addr_len;
   /** deadline for connection */
   gpr_timespec deadline;
   /** channel arguments (to be passed to transport) */
